# packages/api-contract/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "details",
          "request_id"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "LoginRequest": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "LogoutResponse": {
        "additionalProperties": false,
        "properties": {
          "authenticated": {
            "const": false,
            "title": "Authenticated",
            "type": "boolean"
          }
        },
        "required": [
          "authenticated"
        ],
        "title": "LogoutResponse",
        "type": "object"
      },
...
      "RegisterRequest": {
        "additionalProperties": false,
        "properties": {
          "display_name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 256,
            "minLength": 12,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password",
          "display_name"
        ],
        "title": "RegisterRequest",
        "type": "object"
      },
...
      "SessionResponse": {
        "additionalProperties": false,
        "properties": {
          "authenticated": {
            "const": true,
            "title": "Authenticated",
            "type": "boolean"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          }
        },
        "required": [
          "authenticated",
          "user"
        ],
        "title": "SessionResponse",
        "type": "object"
      },
      "UserResponse": {
        "additionalProperties": false,
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "workspace_id": {
            "format": "uuid",
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "display_name",
          "workspace_id"
        ],
        "title": "UserResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "SessionCookie": {
        "description": "Opaque authenticated Trax OS session cookie.",
        "in": "cookie",
        "name": "trax_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Versioned foundation contract. Pydantic wire models and FastAPI route metadata are the canonical authored source for the V1 HTTP contract.",
    "summary": "Public Trax OS API foundation",
    "title": "Trax OS API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login_route_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required or credentials invalid",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authenticated request is not permitted or CSRF check failed",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Version or uniqueness conflict",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server error",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Login Route"
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "logout_route_api_v1_auth_logout_post",
        "parameters": [
          {
            "description": "Double-submit token matching the trax_csrf cookie.",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required or credentials invalid",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authenticated request is not permitted or CSRF check failed",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Version or uniqueness conflict",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server error",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Logout Route"
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "operationId": "register_route_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required or credentials invalid",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authenticated request is not permitted or CSRF check failed",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Version or uniqueness conflict",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server error",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Register Route"
      }
    },
    "/api/v1/auth/session": {
      "get": {
        "operationId": "session_route_api_v1_auth_session_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required or credentials invalid",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authenticated request is not permitted or CSRF check failed",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Version or uniqueness conflict",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server error",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Session Route"
      }
    },
...
  }
}